Update README.md
To build apfs-fuse, I also had to install the packages from build-essential. I would always receive an error that "gcc-c++" could not be found (this was on Ubuntu 22.04 and Xubuntu 22.04).

Amazing, amazing software by the way. Thank you so much for creating this, it has helped me so much with moving my encrypted apfs drives to ext4. Thank you! If there is a way for me to donate money, please let me know!